Take into account cfs quota for recommended_threads call.
The way we convert quota to cpu count is by modelling full time on one cfs_period as one cpu. Thus 100% is 1 cpu, 200% is 2 and so on. We return 1 cpu if the quota is <100% as well. This is how systemd models CPUQuota as well as lxcfs.

We use this cpu count to determine how many logical cpus are available for recommended_threads.

Signed-off-by: Sebastien Dabdoub <sebastien.dabdoub@gmail.com>