Improve testing by actually creating cgroups.
The current set of bats tests create and mount a tempdir and use that as the rootfs for the example job. This doesnt work with sched_getaffinity, so I made them create a temp cgroup and run the example program in there.

In fact, the previous tests didnt even check the output of the example program and thus were always passing.

Signed-off-by: Sebastien Dabdoub <sebastien.dabdoub@gmail.com>